organize cache file models
