+++ lib/http_cache_stream.dart
@@ -16,10 +16,11 @@ library;
 export 'src/cache_manager/http_cache_manager.dart';
 export 'src/cache_server/http_cache_server.dart';
 export 'src/cache_stream/http_cache_stream.dart';
+export 'src/models/cache_files/cache_file_type.dart';
+export 'src/models/cache_files/cache_files.dart';
 export 'src/models/config/cache_config.dart';
 export 'src/models/config/global_cache_config.dart';
 export 'src/models/config/stream_cache_config.dart';
-export 'src/models/metadata/cache_files.dart';
 export 'src/models/metadata/cache_metadata.dart';
 export 'src/models/metadata/cached_response_headers.dart';
 export 'src/models/stream_requests/int_range.dart';
