+++ lib/src/models/metadata/cache_metadata.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:http_cache_stream/src/models/metadata/cache_files.dart';
+import 'package:http_cache_stream/src/models/cache_files/cache_files.dart';
 import 'package:http_cache_stream/src/models/metadata/cached_response_headers.dart';
 
 /// Metadata for a cached file.
